Better collision
-Collision more accurate, however it is still only on top
-tiles.xml now includes a variance (position) variable which is unused
and also it contains a attribute (solid liquid gas) which helps with
collision.
-Collision 3d uses doubles vs ints
-More loading text in console

+++ include/tools.h
@@ -27,7 +27,7 @@ extern bool collisionRight(int xMin1, int xMax1, int xMin2, int xMax2);
 extern bool collisionLeft(int xMin1, int xMax1, int xMin2, int xMax2);
 
 // 3D Collision
-extern bool collision3d(int x1, int width1, int x2, int width2, int y1, int height1, int y2, int height2, int z1, int depth1, int z2, int depth2);
+extern bool collision3d(double x1, double width1, double x2, double width2, double y1, double height1, double y2, double height2, double z1, double depth1, double z2, double depth2);
 
 // Mouse position including resolution difference
 extern int mouseX();
